+++ image_classification/CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 0.0.14 (????-??-??)
 -------------------
 
-- ...
+- added `--replace_chars` option to `retrain.py` to better manage what characters get replaced with blanks in labels
 
 
 0.0.13 (2021-04-16)